[DOP-9787] Avoid implicit updates of MemoryHWMStore data
